% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bldR.R
\docType{package}
\name{bldR}
\alias{bldR}
\alias{bldR-package}
\title{bldR}
\description{
\code{bldR} is a package for modelling bilingual-led divergence of lexica.
}
\details{
bldR provides three R6 classes:
\code{\link{TensorAgent}},
\code{\link{PopulationSimulation}}, and
\code{\link{L2017}}.

Each are described on their own pages.
}
\seealso{
Most uptodate version at http://bld.markellison.net/
}
