% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bldR.R
\docType{data}
\name{EM2017_ParameterFit}
\alias{EM2017_ParameterFit}
\title{EM2017_ParameterFit}
\format{A data frame with 35968726 rows and 9 columns. Of these, 3 columns are for internal use and can be ignored. The others fields are the following:
\describe{
  \item{Form}{The form returned used (one doppel per stimulus, and abstract values indicating non-doppels used in English and Dutch).}
  \item{Meaning}{Names for th 41 distinct meanings tested in the experiment. These take the form of "S"+<stimulus-number+"_"+(allcaps)<meaning>.}
  \item{Language}{The language the simulated participant is trying to express their form in: either English or Dutch. In the experiment reported in EM,
articipants were only aiming to produce forms in English, so the values for \code{Language="Dutch"} are not used in parameter-fitting.}
  \item{lm}{Language mode varying from 0.00 to 1.00 in increments of 0.01.}
  \item{ml}{Monitoring level varying from 0.00 to 1.00 in increments of 0.01.}
\item{p}{The conditional probability of the form given the parameter values - basically, how well did the model do at predicting the frequency with which this form would be used by bilinguals.}
}}
\description{
This data set contains the fit of the cognitive model described in Ellison & Miceli (2017 - Language 93(2):255-287) - hereafter EM -
to the data returned by the experiment also described in the paper.
}
\references{
\url{https://github.com/tyrannomark/bldR}
}
\author{
T. Mark Ellison \email{m.ellison@anu.edu.au}
}
\keyword{data,}
\keyword{experiment}
