% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bldR.R
\docType{data}
\name{EM2017_Responses}
\alias{EM2017_Responses}
\title{EM2017_Responses}
\format{A data frame with 2049 rows and 5 columns. The columns are as follows:
\describe{
\item{Condition}{Either "Monolingual" or "Bilingual" indicating the linguistic background of the participant.}
\item{ParticipantId}{The identifying code for the participant.}
\item{StimulusId}{The identifying code for the stimulus.}
\item{Response}{The response given by this participant to this stimulus.}
\item{IsDoppel}{A numeric value with 1 if the response is a doppel, and 0 if not.}
}}
\description{
This data set contains the responses returned in the experiment described in Ellison & Miceli (2017 - Language 93(2):255-287).
}
\references{
\url{http://bld.markellison.net/}
}
\author{
T. Mark Ellison \email{m.ellison@anu.edu.au}
}
\keyword{data,}
\keyword{experiment}
