% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bldR.R
\docType{data}
\name{EM2017_Stimuli}
\alias{EM2017_Stimuli}
\title{EM2017_Stimuli}
\format{A data frame with 41 rows and 7 columns. The columns are as follows:
\describe{
\item{StimulusId}{An identifying number associated with this particular stimulus.}
\item{DutchDoppel}{The Dutch half of the doppel which is a potential response to the stimulus.}
\item{EnglishDoppel}{The English half of the doppel which is a potential response to the stimulus. This is a potential response to the stimulus.}
\item{EnglishNonDoppelExample}{An example word which could also be a response to the stimulus, but which has no doppel in Dutch.}
\item{DutchContext}{This is the paragraph immediately preceding the sentence frame, tightening the semantic context and (when used with EnglishFrame) pushing the participants into bilingual mode.}
\item{EnglishContext}{This is the paragraph immediately preceding the sentence frame used with monolingual English speakers.}
\item{EnglishFrame}{This is the sentence frame (with a non-initial gap) used to elicit a response from the participants.}
}}
\description{
This data set contains the stimuli used in the experiment described in
Ellison & Miceli (2017 - Language 93(2):255-287).
}
\references{
\url{https://github.com/tyrannomark/bldR}
}
\author{
T. Mark Ellison \email{m.ellison@anu.edu.au}
}
\keyword{data,}
\keyword{experiment}
