% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_directed_cycles.R
\name{get_directed_cycles}
\alias{get_directed_cycles}
\title{Find unique cycles in directed graph}
\usage{
get_directed_cycles(G)
}
\arguments{
\item{G}{An igraph graph.
Must be directed.}
}
\value{
A list of directed cycles.
}
\description{
\code{get_directed_cycles} lists the cycles in a directed graph \code{G}.
}
