% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mixing_matrix.R
\name{get_mixing_matrix}
\alias{get_mixing_matrix}
\title{Compute mixing matrix of a graph}
\usage{
get_mixing_matrix(G, att, counts = FALSE)
}
\arguments{
\item{G}{An igraph graph.}

\item{att}{An attribute name.}

\item{counts}{Whether to return edge counts rather than proportions.
Defaults to FALSE.
If TRUE then returns matrix defined by Equation (1) in Newman (2003).}
}
\value{
A matrix.
}
\description{
\code{get_mixing_matrix} returns the mixing matrix of a graph \code{G} based
  on a node attribute \code{att}.
}
\references{
Newman, M. E. J, (2003).
Mixing patterns in networks.
\emph{Physical Review E}
67:026126.
\doi{10.1103/PhysRevE.67.026126}
}
