% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_many2one.R
\name{is_many2one}
\alias{is_many2one}
\title{Test for many-to-one correspondence between vectors}
\usage{
is_many2one(x, y, na.rm = T)
}
\arguments{
\item{x, y}{Vectors of equal length.}

\item{na.rm}{Should NA values be ignored?
Default is TRUE.}
}
\description{
\code{is_many2one} tests whether there is a many-to-one map from the values
in a vector \code{x} to the corresponding values in a vector \code{y}.
}
