% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2017.R
\docType{class}
\name{L2017}
\alias{L2017}
\title{L2017}
\format{\code{\link{R6Class}} object.}
\usage{
L2017
}
\value{
An object of class L2017.

Object of \code{\link{L2017}} with methods for creating graphs for Ellison & Miceli (2017) published in Language 93(2):255-287.
}
\description{
This class contains methods for creating the graphs in Ellison & Miceli (2017 - Language 93(2):255-287).
These graphs are based on simulations from the classes \code{TensorAgent} and \code{PopulationSimulation}.
There is a method corresponding to each graph, which creates the graph in PNG format and writes it to the
current working directory.
}
\details{
A further method uses the Student t test to evaluate the significance of the difference in distribution
between the monolingual and bilingual results from the experiment.
}
\section{Methods}{

\describe{
  \item{Documentation}{Presents the meanings and uses of \code{L2017}'s methods.}
  \item{\code{$new()}}{Creates a new, empty lexicon object.}
  \item{\code{$draw_1ab(graph_number,language,side)}}{Draws graphs 1a and 1b, depending on arguments (see example).}
  \item{\code{$t_test()}}{Performs the Student t test comparing doppel rates between monolinguals and bilinguals in the experimental results.}
  \item{\code{$draw_2()}}{Draws graph 2.}
  \item{\code{$draw_3a()}}{Draws graph 3a.}
  \item{\code{$draw_3b()}}{Draws graph 3b.}
  \item{\code{$draw_4a()}}{Draws graph 4a.}
  \item{\code{$draw_4b(graphnum="4b")}}{Draws graph 4b.}
  \item{\code{$draw_4c()}}{Draws graph 4c.}
  \item{\code{$draw_6()}}{Draws graph 6.}
  \item{\code{$draw_7a()}}{Draws graph 7a.}
  \item{\code{$draw_7b()}}{Draws graph 7b.}
  \item{\code{$draw_8()}}{Draws graph 8.}
  \item{\code{$draw_9()}}{Draws graph 9.}
  \item{\code{$draw_10a(rx=0.2,ry=0.2,xc=0.5,yc=0.5,A=100,B=100,AB=100,linetypes=c("solid","dashed"),linecolours=c("red","green"),fillcolours=c("red","green"))}}{Draws graph 10a - graphs 10b, 10c build on this function, changing the default parameter values.}
  \item{\code{$draw_10b()}}{Draws graph 10b.}
  \item{\code{$draw_10c()}}{Draws graph 10c.}
  \item{\code{$draw_all()}}{Draws graphs all graphs from the paper.}
}
}

\examples{
library(bldR)
self$t_test();
self$draw_1ab("1a","SG","left");
self$draw_1ab("1b","E","right");
self$draw_2();
self$draw_3a();
self$draw_3b();
self$draw_4a();
self$draw_4b();
self$draw_4c();
self$draw_6();
self$draw_7a();
self$draw_7b();
self$draw_9();
self$draw_10a();
self$draw_10b();
self$draw_10c();
}
\keyword{datasets}
