% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2tbl.R
\name{mat2tbl}
\alias{mat2tbl}
\title{Convert matrices to tibbles}
\usage{
mat2tbl(mat, col.names = c("row", "col", "value"))
}
\arguments{
\item{mat}{Matrix to convert.}

\item{col.names}{Tibble column names.
Defaults to \code{c("row", "col", "value")}.}
}
\description{
\code{mat2tbl} converts a matrix \code{mat} to a tibble.
}
\seealso{
\code{\link{mat2df}}
}
