% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min2hms.R
\name{min2hms}
\alias{min2hms}
\title{Convert decimal number of minutes to h:mm:ss representation}
\usage{
min2hms(x, nearest = T)
}
\arguments{
\item{x}{Decimal number of minutes to be converted.}

\item{nearest}{Whether x should be rounded to the nearest second.
Otherwise x is rounded down.
Default is TRUE.}
}
\value{
A character scalar.
}
\description{
\code{min2hms} converts a decimal number of minutes to a string
  representing the time in h:mm:ss format.
  h may exceed 23 and is omitted if h = 0.
}
\examples{
min2hms(1624.145)  # 27:04:09

}
