% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_coocc.R
\name{normalise_coocc}
\alias{normalise_coocc}
\title{Normalise co-occurrence matrix}
\usage{
normalise_coocc(mat, method = "jaccard")
}
\arguments{
\item{mat}{Co-occurrence matrix.
Must be square.}

\item{method}{Normalisation method.
The default is "jaccard".
Other methods include "cosine" (same as "ochiai") and "dice".}
}
\description{
\code{normalise_coocc} computes the normalisation of a co-occurrence
matrix \code{mat}.
}
\references{
van Eck, N.J. and L. Waltman (2009).
How to normalize cooccurrence data?
An analysis of some well‐known similarity measures.
\emph{Journal of the American Society for Information Science and Technology}
60(8):1635-1651.
}
