% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null2na.R
\name{null2na}
\alias{null2na}
\title{Replace NULLs with NAs}
\usage{
null2na(x)
}
\arguments{
\item{x}{Input object.}
}
\description{
\code{null2na} replaces NULLs with NAs.
}
\examples{
sapply(list(1, 2, NULL, 4), null2na)

}
