% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ppm.R
\name{sample_ppm}
\alias{sample_ppm}
\title{Sample planted partition model}
\usage{
sample_ppm(memb, p, q, seed = NULL)
}
\arguments{
\item{memb}{Vector of community memberships.}

\item{p}{Probability that vertices in same community are adjacent.}

\item{q}{Probability that vertices in different communities are adjacent.}

\item{seed}{Initial seed for edge randomisation (optional).}
}
\value{
An igraph graph.
}
\description{
\code{sample_ppm} generates a random graph using the planted partition model,
  which is a special case of the stochastic block model.
}
