% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_wiener.R
\name{sample_wiener}
\alias{sample_wiener}
\title{Simulate a Wiener process}
\usage{
sample_wiener(mu = 0, sigma = 1, t_max = 1, dt = 0.001)
}
\arguments{
\item{mu}{Drift parameter. Default is 0.}

\item{sigma}{Variance parameter. Default is 1.}

\item{t_max}{Time horizon. Default is 1.}

\item{dt}{Time step. Default is 0.001.}
}
\value{
Data frame with columns
\describe{
\item{t}{Time}
\item{W}{Simulated value}
}
}
\description{
\code{sample_wiener} generates one realization of a Wiener process.
}
