% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_session_info.R
\name{save_session_info}
\alias{save_session_info}
\title{Save session information}
\usage{
save_session_info(path = "session.log")
}
\arguments{
\item{path}{Path of written file.}
}
\description{
\code{save_session_info} writes the output of
\code{sessioninfo::session_info()} to file.
}
