% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_cols.R
\name{sort_cols}
\alias{sort_cols}
\title{Alphabetise data frame and tibble columns}
\usage{
sort_cols(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{...}{(Optional) unquoted expressions determining left-most columns.
See \code{dplyr::select} for syntax requirements.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
\code{sort_cols} alphabetises the columns of a data frame or tibble.
}
\examples{
sort_cols(mtcars, weight = wt)  # left-most column will be "weight"

}
