% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_many2one.R
\name{assert_many2one}
\alias{assert_many2one}
\title{Assert that data contain many-to-one map between columns}
\usage{
assert_many2one(.data, x, y)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{x, y}{Columns intended to share a many-to-one map.}
}
\description{
\code{assert_many2one} checks whether there is a many-to-one map between
two columns \code{x} and \code{y} of a data frame \code{.data}, raising an
error if not and returning \code{.data} otherwise.
}
