% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationSimulation.R
\docType{class}
\name{PopulationSimulation}
\alias{PopulationSimulation}
\title{PopulationSimulation}
\format{\code{\link{R6Class}} object.}
\usage{
PopulationSimulation
}
\value{
An object of class PopulationSimulation.

Object of \code{\link{PopulationSimulation}} with methods for drawing graphs according suitable for final version of the paper.
}
\description{
This R6 class defines functions for running agent-based simulations described in Ellison & Miceli (2017 - Language 93(2):255-287) - hereafter EM.
These simulations use the cognitive model
defined by the class TensorAgent in this package. The modelling works by setting the parameters using
auxiliary methods, then calling a \code{$simulate()} method to execute the simulation.
The simulation models a two-language community, with 2 monolingual populations and a bilingual community.
The sizes of each of these communities can be set by parameter. The language mode and monitoring level are
set globally in the simulation, applying to all individuals in the bilingual community (they have no effect
on the monolingual communities).
}
\details{
Graphs 10a,b,c in EM are based on simulations run using this class.
}
\examples{
library(bldR)
ps <- PopulationSimulation$new();
ps$setPopulationStructure(20,20,20);
ps$setNumberOfGenerations(10);
ps$setMonitoringLevel(1.0);
ps$setLanguageMode(0.54);
ps$setSamplesPerAgent(100);
ps$setLexicon();
ps$simulate(exact=FALSE);

}
\keyword{datasets}
