% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec2hms.R
\name{sec2hms}
\alias{sec2hms}
\title{Convert decimal number of seconds to h:mm:ss representation}
\usage{
sec2hms(x, nearest = T)
}
\arguments{
\item{x}{Decimal number of seconds to be converted.}

\item{nearest}{Whether x should be rounded to the nearest second.
Otherwise x is rounded down.
Default is TRUE.}
}
\value{
A character scalar.
}
\description{
\code{sec2hms} converts a decimal number of seconds to a string
  representing the time in h:mm:ss format.
  h may exceed 23 and is omitted if h = 0.
}
\examples{
sec2hms(97448.7)  # 27:04:09

}
